Bugfix in purgeGraph method (syntax error in SPARQL query)
branching strategy changed (#12)
 - wasDerivedFrom: identifies the revision on which the branch based
 - wasAssociatedWith: identifies the user who branches
 - title: the branch message
 - atTime: the time stamp of the branching
After the creation of the branch the branch has no rmo:references. This
will be generated after the first commit to the branch. Therefore the
user has to use the branch label as revision name in the commit.


missing negation added (#14)


existence check for branches before creation (#13)


some little bugfixes